BFS.py
algorithm to find shortest distance to some node s from all other nodes in a graph. Input the number of queries on a line, then for the number of queries do the following. Input the number of nodes followed by a space and number of edges on their own line, then for the number of edges make a line with two numbers separated by a space which means there is an edge between those nodes, and then put a number for the id of the starting node on the last line. This algorithm is not quite fast enough yet. However, it is correct.